wednesday check-in
- add optional status parameter sensor update
- add data-sensing-lab icon
- minor arduino cleanup

+++ things/examples/arduino/GroveAQ/GroveAQ.ino
@@ -11,7 +11,7 @@
 #include "AirQuality.h"
 
 
-int requestID = 1;
+unsigned long requestID = 1;
 unsigned long next_heartbeat = 0;
