+++ things/examples/arduino/WaterSensor/WaterSensor.ino
@@ -9,7 +9,7 @@
 #include <SPI.h>
 
 
-int requestID = 1;
+unsigned long requestID = 1;
 unsigned long next_heartbeat = 0;
